♻️ refactor(User): User 데이터 파편화 방지를 위한 Service 로직 통일 (#147)
