+++ src/main/java/com/sparta/devquiz/domain/board/service/BoardService.java
@@ -55,12 +55,6 @@ public class BoardService {
         return BoardDetailsResponse.of(boards);
     }
 
-    public List<BoardDetailsResponse> getBoardListByUserId(Long userId) {
-        List<Board> boards = boardRepository.findAllByUserIdAndIsDeletedFalse(userId);
-
-        return BoardDetailsResponse.of(boards);
-    }
-
     @Transactional
     public void updateBoard(Long boardId, BoardUpdateRequest request, User user) {
         Board board = boardRepository.findBoardByIdOrElseThrow(boardId);
