+++ src/main/java/com/sparta/devquiz/domain/quiz/service/QuizService.java
@@ -8,7 +8,11 @@ import com.sparta.devquiz.domain.coin.service.CoinService;
 import com.sparta.devquiz.domain.quiz.dto.request.QuizAnswerSubmitRequest;
 import com.sparta.devquiz.domain.quiz.dto.request.QuizCreateRequest;
 import com.sparta.devquiz.domain.quiz.dto.request.QuizUpdateRequest;
-import com.sparta.devquiz.domain.quiz.dto.response.*;
+import com.sparta.devquiz.domain.quiz.dto.response.QuizDetailInfoResponse;
+import com.sparta.devquiz.domain.quiz.dto.response.QuizPassResponse;
+import com.sparta.devquiz.domain.quiz.dto.response.QuizQueryResponse;
+import com.sparta.devquiz.domain.quiz.dto.response.QuizRandomResponse;
+import com.sparta.devquiz.domain.quiz.dto.response.QuizResultResponse;
 import com.sparta.devquiz.domain.quiz.entity.Quiz;
 import com.sparta.devquiz.domain.quiz.entity.QuizChoice;
 import com.sparta.devquiz.domain.quiz.entity.UserQuiz;
