+++ src/test/java/com/sparta/devquiz/domain/coin/service/CoinServiceTest.java
@@ -1,19 +1,10 @@
 package com.sparta.devquiz.domain.coin.service;
 
-import com.sparta.devquiz.domain.coin.dto.request.CoinSaveRequest;
-import com.sparta.devquiz.domain.coin.dto.request.CoinUseRequest;
-import com.sparta.devquiz.domain.coin.dto.response.CoinGetInfoResponse;
-import com.sparta.devquiz.domain.coin.dto.response.CoinUseResponse;
-import com.sparta.devquiz.domain.coin.entity.Coin;
-import com.sparta.devquiz.domain.coin.enums.CoinContent;
-import com.sparta.devquiz.domain.coin.exception.CoinCustomException;
 import com.sparta.devquiz.domain.coin.repository.CoinRepository;
 import com.sparta.devquiz.domain.user.entity.User;
 import com.sparta.devquiz.domain.user.repository.UserRepository;
-import com.sparta.devquiz.domain.user.service.command.UserService;
+import com.sparta.devquiz.domain.user.service.UserService;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
@@ -21,13 +12,9 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.test.context.ActiveProfiles;
 
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.Mockito.lenient;
-import static org.mockito.Mockito.when;
 
 @ActiveProfiles("test")
 @ExtendWith(MockitoExtension.class)
